cleanup and shape 3 changes
cleaned up some comments and added some new comments

also changed the n() function to return length of data
reglardless of shape. Before, we were returning length
of first dimension from n() and we needed to use nz()
to get the length of a shape 3 noun. Now we can use n()
to get the data length of any noun. the verb ct() now
houses to logic to return the length of first dimension of
a shape 3 noun.

also fixed a but in one of the concatenation verbs that
was writing offsets into the length vectors.