no segfault on verb projection
before, expressions like + and 1+2+ or 1++ would fail
because the evaluator assumes all data is available for
the verbs. now, partial evaluation is supported by returning
a type 4 that acts like a type 0 but represents missing data.

evaluation functions e emv and edv can sense when a type 4 is
returned and handles it as required.

next is to apply data to partially evaluated programs to run them
that is currently still broken.