tag switched to low bits
instead of using the top 16 bits (of which only the low 8 were used)
we use the empty bits on the low side of the pointer. All pointers
are 8-byte aligned so we use 3 bits.

0 pointer
1 ref
2 verb
3 int
4 partial eval

shape and other properties are only ever relevant for heap allocated
objects.

AH and Z are now 8 byte aligned.

Pointers now point at the head instead of at the data.