state machine lexer and global hash table
references,symbols,multi digit numbers are now supported in the lexer.
basically each of those types are treated as a sort of integer but with
a different radix depending on the legal characters.

also, a global hash table was introduced so that we can store more than
a-z. The hash table utilizes fibonacci multiplicative hashing and open
adressing with linear probing.