major overhaul for dictionaries
added type 5 for hash. hash is a vector of dwords that are indices into
key/value arrays which are stored as siblings with the hash. a
dictionary is an (h;k;v) list with type 0 shape 2. In the future I would
like to support typed dictionaries that support storage of just a single
type.

getters and setters were changed to be explicit about supporting only
shape 1 and scalar expansion was removed from the qi function.

the hash table uses 0 as empty, which implies that it stores 1 based
indices. you need to subtract 1 from the index received from the hash
table in order to properly index into the kv arrays in the dictionary.

A single global dictionary is allocated right now. This is building
towards scopes and dictionaries allocated in temp space.

right now a dictionaries capacity is hard coded to 64 but it should be a
function of the n passed in.

Also, a notion of capacity was added to the header so that we can
represent growth of vectors but allocate space in advance for the
growth. for example a hash is size 0 but when we write a new key we
increment the size by 1. if n==c we should grow the allocation but that
is left unimplemented.