introduce scopes
{ opens a new scope
} closes the scope

scopes are bump allocated.
at the end of an evaluation, all data in the root scope is rehashed into
the global dictionary G.

There are various bugs but getting this in as a first cut.
known issues:
1. popping scopes seems to corrupt lists in higher scopes. e.g.
a:!10;b:{...};a prints an atom type 0 which is corrupted.
2. concatenation of vectors is broken
3. need to end dictionaries fully initialized on a single line with ;}
for some reason just } causes issues.
4. it is possible to create cycles in dictionaries, something to do with
the lifecycle of when } happens

probably others. but it is pretty cool as is.

also fixed a bug in the allocator which was not memsetting the right
amount of bytes, due to being passed in Q sized units instead of bytes.