adverbs!!!
the way adverbs work is that they modify verbs by encoding themselves
onto the verbs. It works in a similar fashion to how numbers and
references are parsed. We use the count of adverbs as a base, the count
of verbs as the base of the starting digit and then multiply the verb by
the adverb list size and add the adverb code. This operation can be
undone deterministically.

The adverb is chosen contextually right now. if we are in edv then \/'
are the dyadic options eachleft eachright and eachdyad

in emv the adverbs are scan over and eachmonad

in order to enable this, the verb itself is passed as a parameter into
verb functions so that we can decode an adverb layer by layer.